refactor(v0.79.0): bash_tool.py core/cli → core/agent 이전 (의존성 정리 A2, 5/5)
새 사이클 A2. v0.52에서 deferred됐던 33개 ignore_imports 중 추가 2개 정리.

이전: core/cli/bash_tool.py (162L) → core/agent/bash_tool.py
- HITL-gated shell 실행 도구 — agentic loop 안에서만 사용
- 호출자가 ToolExecutor (core.agent) 뿐. CLI 의존성 0
- v0.22 historical artifact였음

호출자 업데이트 (3 파일, 7곳):
- core/agent/tool_executor/executor.py:20
- tests/test_bash_tool.py:8,140,146
- tests/test_redaction.py:80,93,106 (lazy imports)

pyproject.toml 변경:
- ignore_imports 2 entries 제거 (Agent stays pure + Server may host agent)
- per-file-ignores S602 path rename (core/cli → core/agent)

31 → 29 ignore_imports 남음.

Quality Gates:
- ruff/ruff-format/mypy 0 errors (329 source files)
- pytest 4344 passed, 1 skipped, 24 deselected — 변동 없음
- import-linter 4/4 contracts 깨짐 0
- E2E Cowboy Bebop A (68.4) 변동 없음

5 PR 진척:
- A1 redaction → utils ✅ v0.78.0
- A2 (this) bash_tool → agent ✅ v0.79.0
- A3 project_detect → utils
- A4 session_checkpoint + transcript → runtime_state/
- A5 startup → lifecycle
