+++ core/agent/bash_tool.py
